requirements.txt
pandas==2.2.2
requests==2.32.3
scikit-learn==1.5.2
numpy==2.0.2
fastapi[standard]==0.115.6
pydantic==2.10.5
nltk==3.9.1
uvicorn==0.34.0
transformers==4.48.1
spacy==3.8.4
python-dotenv==1.0.1
pydantic-settings==2.10.1
pymongo==4.6.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
bson==0.5.10
datasets==4.0.0
torch==2.8.0